Move test files next to code being tested
