+++ go_api/handler/server/user_test.go
@@ -1,4 +1,4 @@
-package tests
+package server
 
 import (
 	"bytes"
@@ -10,7 +10,6 @@ import (
 
 	"the-drink-almanac-api/apperrors"
 	"the-drink-almanac-api/dto"
-	"the-drink-almanac-api/handler/server"
 	"the-drink-almanac-api/mocks"
 	"the-drink-almanac-api/model"
 
@@ -62,7 +61,7 @@ func TestFindUser(t *testing.T) {
 				mockUserService.On("FindUser", d.userId).Return(d.returnedUser, d.returnedError)
 			}
 			mockAuthService := mocks.NewAuthService(t)
-			userHandler := server.NewUserHandler(mockUserService, mockAuthService)
+			userHandler := NewUserHandler(mockUserService, mockAuthService)
 
 			rr := httptest.NewRecorder()
 			request, err := http.NewRequest(http.MethodGet, "/user", nil)
@@ -191,7 +190,7 @@ func TestCreateNewUser(t *testing.T) {
 				mockUserService.On("CreateNewUser", d.username, d.password).Return(d.returnedUser, d.returnedError)
 			}
 			mockAuthService := mocks.NewAuthService(t)
-			userHandler := server.NewUserHandler(mockUserService, mockAuthService)
+			userHandler := NewUserHandler(mockUserService, mockAuthService)
 
 			rr := httptest.NewRecorder()
 			request, err := http.NewRequest(http.MethodPost, "/user", bytes.NewBuffer(d.requestBody))
@@ -253,7 +252,7 @@ func TestDeleteUser(t *testing.T) {
 			if d.userId != "" {
 				mockUserService.On("DeleteUser", d.userId).Return(d.returnedError)
 			}
-			userHandler := server.NewUserHandler(mockUserService, mockAuthService)
+			userHandler := NewUserHandler(mockUserService, mockAuthService)
 
 			rr := httptest.NewRecorder()
 			request, err := http.NewRequest(http.MethodDelete, "/user", nil)
@@ -376,7 +375,7 @@ func TestLogin(t *testing.T) {
 			if d.shouldReturnToken {
 				mockAuthService.On("CreateNewToken", d.returnedUser.Id, mock.Anything).Return("testToken", nil)
 			}
-			userHandler := server.NewUserHandler(mockUserService, mockAuthService)
+			userHandler := NewUserHandler(mockUserService, mockAuthService)
 
 			rr := httptest.NewRecorder()
 			request, err := http.NewRequest(http.MethodPost, "/user/login", bytes.NewBuffer(d.requestBody))
