+++ go_api/handler/server/util.go
@@ -1,4 +1,4 @@
-package tests
+package server
 
 import "github.com/gin-gonic/gin"
