+++ go_api/service/auth_test.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	"testing"
-	"the-drink-almanac-api/service"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -39,7 +38,7 @@ func TestJwtAuthService(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			authService := service.NewJwtAuthService("testToken")
+			authService := NewJwtAuthService("testToken")
 
 			var tokenString string
 			if !tt.useFakeToken {
